refactor(useTocHighlight): 하이라이트 기능 훅으로 분리
